% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{colon_heatmap_grid}
\alias{colon_heatmap_grid}
\title{Colon Heatmap Grid}
\usage{
colon_heatmap_grid(grid_res = 500, variant = "standard")
}
\arguments{
\item{grid_res}{The resolution of the grid, which determines how many points are generated.
A higher resolution will create a denser grid, which can lead to
smoother heatmaps but may also increase computation time.}

\item{variant}{A character string specifying the variant of the colon
blueprint to use.}
}
\value{
An sf object containing the points that lie within the colon geometry,
which can be used for heatmap generation or other spatial analyses.
}
\description{
This function creates a grid of points within the colon geometry,
which can be used to generate heatmaps or other spatial visualizations.
The grid is created by intersecting a dense grid of points with the colon geometry,
resulting in a set of points that lie within the colon structure.
}
\details{
The function first creates a dense grid of points based on the specified resolution,
then intersects this grid with the colon geometry to retain only the points that
lie within the colon. The resulting sf object contains the coordinates of the points
that can be used for further analysis or visualization.
}
