% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{colon_sf}
\alias{colon_sf}
\title{Colon Geometry as sf Object}
\usage{
colon_sf(variant = "standard")
}
\arguments{
\item{variant}{A character string specifying the variant of the colon
blueprint to use.}
}
\value{
An sf object with polygon geometries for each part of the colon.
}
\description{
This function converts the colon blueprint into an sf object,
which can be used for spatial operations and visualizations. It groups the
blueprint by colon parts and creates polygon geometries for each part.
}
\details{
The resulting sf object contains the geometry of the colon parts,
which can be used for various spatial analyses or visualizations, such as
creating heatmaps or overlaying additional data on the colon structure.
}
