% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{ggcolonviz-package}
\alias{ggcolonviz}
\alias{ggcolonviz-package}
\title{\code{ggcolonviz} Package}
\description{
\code{ggcolonviz} implements a ggplot2 extension for visualizing the colon.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Sanofi-Public/ggcolonviz}
  \item Report bugs at \url{https://github.com/Sanofi-Public/ggcolonviz/issues}
}

}
\author{
\strong{Maintainer}: Yoni Sidi \email{yoni.sidi@sanofi.com}

Other contributors:
\itemize{
  \item Sanofi [copyright holder]
}

}
